# shared/openapi.json
{
  "components": {
    "schemas": {
      "DemoModeIn": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "DemoModeIn",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobActionOut": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "ok",
          "status"
        ],
        "title": "JobActionOut",
        "type": "object"
      },
      "JobOut": {
        "properties": {
          "active_gpu_seconds": {
            "title": "Active Gpu Seconds",
            "type": "number"
          },
          "config_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Summary"
          },
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "flops_used": {
            "title": "Flops Used",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "param_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Param Count"
          },
          "participant_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Participant Code"
          },
          "preemptions": {
            "title": "Preemptions",
            "type": "integer"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "samples_seen": {
            "title": "Samples Seen",
            "type": "integer"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "stop_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Reason"
          },
          "tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tier"
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "priority",
          "samples_seen",
          "flops_used",
          "active_gpu_seconds",
          "preemptions",
          "created_at"
        ],
        "title": "JobOut",
        "type": "object"
      },
...
      "OkOut": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "OkOut",
        "type": "object"
      },
...
      "QueueOut": {
        "properties": {
          "budget_flops": {
            "title": "Budget Flops",
            "type": "number"
          },
          "current_job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Job Id"
          },
          "demo_mode": {
            "title": "Demo Mode",
            "type": "boolean"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobOut"
            },
            "title": "Jobs",
            "type": "array"
          },
          "paused": {
            "title": "Paused",
            "type": "boolean"
          }
        },
        "required": [
          "paused",
          "demo_mode",
          "current_job_id",
          "budget_flops",
          "jobs"
        ],
        "title": "QueueOut",
        "type": "object"
      },
...
      "ResetIn": {
        "properties": {
          "confirm": {
            "title": "Confirm",
            "type": "string"
          }
        },
        "required": [
          "confirm"
        ],
        "title": "ResetIn",
        "type": "object"
      },
      "ResetOut": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "snapshot": {
            "title": "Snapshot",
            "type": "string"
          }
        },
        "required": [
          "ok",
          "snapshot"
        ],
        "title": "ResetOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Build Your Own AI backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/demo-mode": {
      "post": {
        "operationId": "demo_mode_admin_demo_mode_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-admin-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoModeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Demo Mode"
      }
    },
    "/admin/export": {
      "get": {
        "operationId": "export_admin_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "pattern": "^(json|csv)$",
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_contact",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Contact",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "x-admin-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export"
      }
    },
    "/admin/jobs/{job_id}": {
      "delete": {
        "operationId": "remove_admin_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-admin-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove"
      }
    },
    "/admin/jobs/{job_id}/kill": {
      "post": {
        "operationId": "kill_admin_jobs__job_id__kill_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-admin-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Kill"
      }
    },
    "/admin/jobs/{job_id}/redo": {
      "post": {
        "operationId": "redo_admin_jobs__job_id__redo_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-admin-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Redo"
      }
    },
    "/admin/logs": {
      "get": {
        "operationId": "logs_admin_logs_get",
        "parameters": [
          {
            "in": "query",
            "name": "tail",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 2000,
              "minimum": 1,
              "title": "Tail",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-admin-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Logs Admin Logs Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logs"
      }
    },
    "/admin/queue": {
      "get": {
        "operationId": "queue_admin_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_removed",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Removed",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "x-admin-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Queue"
      }
    },
    "/admin/queue/pause": {
      "post": {
        "operationId": "pause_admin_queue_pause_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-admin-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pause"
      }
    },
    "/admin/queue/resume": {
      "post": {
        "operationId": "resume_admin_queue_resume_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-admin-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume"
      }
    },
    "/admin/reset": {
      "post": {
        "description": "Nuclear reset: snapshot the database FIRST, then clear queue, results and stored runs.",
        "operationId": "reset_admin_reset_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-admin-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset"
      }
    },
    "/admin/stream": {
      "get": {
        "description": "Live monitor (SSE): one `status` event per second with the current job's curves, GPU stats and a log tail.",
        "operationId": "stream_admin_stream_get",
        "parameters": [
          {
            "description": "stop after N events (tests)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "stop after N events (tests)",
              "title": "Limit"
            }
          },
          {
            "in": "header",
            "name": "x-admin-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream"
      }
    },
...
  }
}